Add more exercise instructions
- Updates exercise instructions for the 'streamlined' version of
  the FIR implementation
- Moves some definitions around to be consistent with the state
  of the previous exercise
