+++ step_100_implement_fir_streamlined/src/main/scala/org/applied/akkastreams/FIRElements.scala
@@ -1,0 +1,19 @@
+package org.applied.akkastreams
+
+import akka.stream.scaladsl.Flow
+
+object FIRElements {
+  object FirInitial {
+    def apply() = {
+      Flow.fromFunction[Double, (Double, Double)] { sample =>
+        (sample, sample)
+      }
+    }
+  }
+
+  object FirSelectOut {
+    def apply() = {
+      Flow.fromFunction[(Double, Double), Double] { case (_, out) => out}
+    }
+  }
+}
